refactor(bump): split io.py, cut complexity, hoist cli imports
Addresses the three quality-scorecard findings from #316.

Closes #317: no block in the bump package now ranks worse than A —
- engine._show_file_changes (B7) → extracted _find_version_lines / _print_line_diff
- engine._build_configuration (B6) → extracted _build_config_overrides
- io._validate_project_exists (B6) → table-driven per-language required files
- prompts._show_interactive_preview (B6) → extracted _confirm helper
- git._handle_branch_checkout (B6) → extracted _checkout_branch

Closes #318: split the 426-LOC bump/io.py into cohesive modules, each <250 LOC —
- models.py    — Language, BumpOptions, parse_language_option (data model)
- io.py        — project-file version reads + validation (now 151 LOC)
- prompts.py   — questionary-backed interactive helpers
- reporting.py — post-bump success/modified-file logging
__init__.py re-exports every symbol so the public import surface is unchanged.

Closes #319: hoisted the function-local bump/rollback imports in cli.py to
module top-level (the modules were already imported there).

Behaviour-preserving: all 493 tests pass at 100% coverage; ty + mypy --strict,
deptry, and interrogate (100%) all green. Test monkeypatch paths updated from
bump.io.qs → bump.prompts.qs to follow the moved interactive helpers.
